Collect the remote-command effects out of MainDesktop
Fourteen command flows — Companion buttons, a phone, a linked instance —
were scattered through 500 lines of unrelated state. They are one concern
and now sit together.

This is the only cluster MainDesktop has. Measured, the tab-content Row has
102 dependencies of which 71 are written to, so extracting it would hand a
new file almost the whole scope and take it back through callbacks: the
lines would move and nothing would decouple. RemoteCommandEffects has 25
dependencies and 3 writes, which is why it is worth having.

The view models are passed rather than wrapped in a callback each, because
the bodies are calls onto them — `nextPictureFlow` is two lines of
picturesViewModel. Callbacks would only move those lines to the call site.
MainDesktop is the documented place where view models are wired top-down;
this stays inside that, and no other file should take one.

MainDesktop 1,945, RemoteCommandEffects 242.
